+++ react-chat/src/modules/ChatList/ChatList.jsx
@@ -1,36 +1,12 @@
-import { useState, useEffect } from "react";
 import { Chat } from "@/components/Chat/Chat";
-import { getAllChats } from "@/services/chat";
 import { Link } from "react-router-dom";
 
-export const ChatList = ({ searchQuery }) => {
-    const [chats, setChats] = useState([]);
-    const [loading, setLoading] = useState(true);
-
-    useEffect(() => {
-        const fetchChats = async () => {
-            try {
-                const response = await getAllChats();
-                setChats(response.results);
-                console.log("Chats fetched:", response.results);
-            } catch (error) {
-                console.error("Error fetching chats:", error);
-            } finally {
-                setLoading(false);
-            }
-        };
-
-        fetchChats();
-    }, []);
+export const ChatList = ({ searchQuery, chats }) => {
+    console.log("Chats", chats);
 
     const filteredChats = chats.filter((chat) => {
         return chat.title.toLowerCase().includes(searchQuery.toLowerCase());
     });
-
-    if (loading) {
-        return <p>Loading chats...</p>;
-    }
-
     return (
         <div>
             {filteredChats.length > 0 ? (
