+++ react-chat/src/modules/Sidebar/Sidebar.jsx
@@ -25,8 +25,6 @@ export const Sidebar = () => {
         fetchUserData();
     }, []);
 
-    //console.log("user data", userData);
-
     const handleToggle = () => {
         setIsOpen(!isOpen);
     };
