+++ react-chat/vite.config.js
@@ -13,8 +13,8 @@ export default defineConfig({
     server: {
         proxy: {
             "/api": {
-                // target: "https://vkedu-fullstack-div2.ru",
-                target: "http://localhost:8080",
+                target: "https://vkedu-fullstack-div2.ru",
+                // target: "http://localhost:8080",
                 changeOrigin: true,
                 rewrite: (path) => path.replace(/^\/api/, ""),
             },
